#137288 - add setup queue, remove old calls to delivery queue
