+++ src/Tribe/API/Timeline.php
@@ -3,7 +3,7 @@
 namespace Tribe\HubSpot\API;
 
 use SevenShores\Hubspot\Factory;
-use Tribe\HubSpot\Process\Delivery_Queue;
+use Tribe\HubSpot\Process\Setup_Queue;
 
 /**
  * Class Timeline
@@ -61,7 +61,7 @@ class Timeline {
 			'type' => 'update_timeline_event_types',
 		];
 
-		$queue = new Delivery_Queue();
+		$queue = new Setup_Queue();
 		$queue->push_to_queue( $hubspot_data );
 		$queue->save();
 		$queue->dispatch();
